imported local from fabric.api fixed pycodestyle and add guard against number < 0

+++ 100-clean_web_static.py
@@ -3,7 +3,7 @@
 import os
 from fabric.api import env
 from fabric.api import run
-
+from fabric.api import local
 env.hosts = ['100.26.20.157', '54.144.137.83']
 
 
@@ -23,12 +23,17 @@ def do_clean(number=0):
         return None
 
     web_files.sort()
+    # print(web_files)
+    if int(number) < 0:
+        return None
     clean_count = count - int(number)
+    # print('clean_count = {}'.format(clean_count))
     if clean_count > 0 and count > int(number) + 1:
-#        if number == 0:
-#            run('rm /data/web_static/releases/{}'.format(web_files[0]))
-#            local('rm versions/{}.tgz'.format(web_files[0]))
-#        else:
+        # if number == 0:
+        #     run('rm /data/web_static/releases/{}'.format(web_files[0]))
+        #     local('rm versions/{}.tgz'.format(web_files[0]))
+        # else:
         for i in range(clean_count):
+            print('/data/web_static/releases/{}'.format(web_files[i]))
             run('rm -r /data/web_static/releases/{}'.format(web_files[i]))
             local('rm versions/{}.tgz'.format(web_files[i]))
